Refactor report modules into package
