+++ src/vigapp/ui/menu_window.py
@@ -104,7 +104,7 @@ from .moment_app import MomentApp
 from .design_window import DesignWindow
 from .shear_design_window import ShearDesignWindow
 from .view3d_window import View3DWindow
-from reporte_flexion_html import generar_reporte_html
+from vigapp.reports.reporte_flexion_html import generar_reporte_html
 
 
 class MenuWindow(QMainWindow):
